+++ hollow/src/main/java/com/netflix/hollow/api/producer/CloseableIncrementalWriteState.java
@@ -1,5 +1,6 @@
 package com.netflix.hollow.api.producer;
 
+import static com.netflix.hollow.api.producer.HollowIncrementalCyclePopulator.AddIfAbsent;
 import static com.netflix.hollow.api.producer.HollowIncrementalCyclePopulator.DELETE_RECORD;
 
 import com.netflix.hollow.core.write.objectmapper.HollowObjectMapper;
@@ -9,7 +10,7 @@ import java.util.concurrent.ConcurrentHashMap;
 
 // @@@ Move into HollowIncrementalCyclePopulator since the event map is a shared resource
 //     HollowIncrementalCyclePopulator constructed from the write state instance
-final class CloseableIncrementalWriteState implements HollowProducer.IncrementalWriteState, AutoCloseable {
+final class CloseableIncrementalWriteState implements HollowProducer.Incremental.IncrementalWriteState, AutoCloseable {
     private final ConcurrentHashMap<RecordPrimaryKey, Object> events;
     private final HollowObjectMapper objectMapper;
     private volatile boolean closed;
@@ -21,39 +22,40 @@ final class CloseableIncrementalWriteState implements HollowProducer.Incremental
         this.objectMapper = objectMapper;
     }
 
-    @Override public void addOrModify(Object o) {
+    @Override
+    public void addOrModify(Object o) {
         ensureNotClosed();
 
-        RecordPrimaryKey key;
-        if (o instanceof FlatRecord) {
-            FlatRecord fr = (FlatRecord) o;
-            key = fr.getRecordPrimaryKey();
-        } else {
-            key = objectMapper.extractPrimaryKey(o);
-        }
-
-        events.put(key, o);
+        events.put(getKey(o), o);
     }
 
-    @Override public void delete(Object o) {
+    @Override
+    public void addIfAbsent(Object o) {
         ensureNotClosed();
 
-        RecordPrimaryKey key;
-        if (o instanceof FlatRecord) {
-            FlatRecord fr = (FlatRecord) o;
-            key = fr.getRecordPrimaryKey();
-        } else {
-            key = objectMapper.extractPrimaryKey(o);
-        }
-
-        delete(key);
+        events.putIfAbsent(getKey(o), new AddIfAbsent(o));
     }
 
-    @Override public void delete(RecordPrimaryKey key) {
+    @Override
+    public void delete(Object o) {
+        delete(getKey(o));
+    }
+
+    @Override
+    public void delete(RecordPrimaryKey key) {
         ensureNotClosed();
 
         // @@@ Deletion is silently ignored if no object exists for the key
         events.put(key, DELETE_RECORD);
+    }
+
+    private RecordPrimaryKey getKey(Object o) {
+        if (o instanceof FlatRecord) {
+            FlatRecord fr = (FlatRecord) o;
+            return fr.getRecordPrimaryKey();
+        } else {
+            return objectMapper.extractPrimaryKey(o);
+        }
     }
 
     private void ensureNotClosed() {
